@@ -3,10 +3,11 @@
     "version": "1.0.0",
     "private": true,
     "dependencies": {
-      "react": "^17.0.2",
-      "react-dom": "^17.0.2",
+      "react": "^18.0.0",
+      "react-dom": "^18.0.0",
       "react-router-dom": "^6.0.0",
-      "react-scripts": "4.0.3"
+      "react-scripts": "4.0.3",
+      "axios": "^0.21.1"
     },
     "scripts": {
       "start": "react-scripts start",
